Implementado exclusão tanto produto quanto imagens e S3

+++ src/modules/Product/services/DeleteProductService.ts
@@ -1,5 +1,8 @@
 import { inject, injectable } from 'tsyringe';
 import IProductRepository from '../repositories/IProductRepository';
+import fs from 'fs';
+import aws from 'aws-sdk';
+const s3 = new aws.S3();
 
 @injectable()
 class DeleteProductService {
@@ -9,7 +12,23 @@ class DeleteProductService {
   ) {}
 
   public async execute(idProduct: string): Promise<void> {
-    await this.productsRepository.deleteById(idProduct);
+    const images = await this.productsRepository.findImagesProduct(idProduct);
+
+    if (images) {
+      if (process.env.STORAGE_TYPE === 's3') {
+        s3.deleteObject(
+          {
+            Bucket: process.env.BUCKET_NAME || '',
+            Key: images[0].key,
+          },
+          err => {
+            console.log(err);
+          },
+        );
+      }
+      await this.productsRepository.deleteImageProduct(images);
+      await this.productsRepository.deleteById(idProduct);
+    }
   }
 }
